Place classes in project structure
