+++ src/Files.App/Terminal/TerminalBridge.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Windows.Foundation.Metadata;
 
-namespace Files.App.UserControls
+namespace Files.App.Terminal
 {
 	/// <summary>
 	/// Code modified from https://github.com/felixse/FluentTerminal
