+++ src/Files.App/UserControls/TerminalView.xaml
