+++ src/Files.App/UserControls/TerminalView.xaml.cs
